requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
openai==0.27.0
Pillow==9.5.0
piexif==1.1.3
tweepy==4.12.0
elevenlabs==0.2.27
gunicorn==21.2.0
gevent==23.9.1 